Refactor testing into multiple files.
Also fixes a bug in which the mapper executable would ignore the
expander argument.
